@@ -1,9 +1,6 @@
 name: log_storage_client
-authors:
-    - "Philipp Perez"
 description: A desktop app to upload and download log files to an S3-based object storage.
-
-author: "Philipp Perez"
+repository: https://github.com/emotionrennteam/log-storage-client/
 
 # The following line prevents the package from being accidentally published to
 # pub.dev using `pub publish`. This is preferred for private packages.
@@ -30,15 +27,15 @@ dependencies:
   draggable_scrollbar: ^0.0.4
   dynamic_color_theme: ^1.1.1
   file_picker: ^2.0.4
-  flutter_staggered_grid_view: ^0.3.2
-  font_awesome_flutter: ^8.8.1
-  get_it: ^5.0.1
+  flutter_staggered_grid_view: ^0.3.4
+  font_awesome_flutter: ^8.12.0
+  get_it: ^6.0.0
   intl: ^0.16.1
-  minio: ^1.1.0
-  path: ^1.7.0
+  minio: ^1.3.0
+  path: ^1.8.0
   shared_preferences: ^0.5.12
   tuple: ^1.0.3
-  url_launcher: ^5.7.2
+  url_launcher: ^6.0.2
 
 dev_dependencies:
   flutter_test:
