@@ -9,12 +9,12 @@ class EmotionDesignButton extends StatelessWidget {
   /// Defaults to the theme's canvas color.
   final Color color;
 
-  /// Encapsulates a [FlatButton] which is styled according to this app's design.
+  /// Encapsulates a [TextButton] which is styled according to this app's design.
   EmotionDesignButton({
     @required this.child,
     @required this.onPressed,
     this.color,
-    this.verticalPadding = 20,
+    this.verticalPadding = 14,
   });
 
   @override
@@ -24,21 +24,27 @@ class EmotionDesignButton extends StatelessWidget {
       color = Theme.of(context).canvasColor;
     }
 
-    return FlatButton(
-      child: this.child,
-      padding: EdgeInsets.symmetric(
-        vertical: this.verticalPadding,
-        horizontal: 30,
-      ),
-      color: color,
-      shape: RoundedRectangleBorder(
-        borderRadius: BorderRadius.circular(BORDER_RADIUS_LARGE),
-        side: BorderSide(
-          color: Color.fromRGBO(40, 40, 40, 1),
-          width: 2,
+    return TextButton(
+      onPressed: this.onPressed,
+      child: Padding(
+        child: this.child,
+        padding: EdgeInsets.symmetric(
+          vertical: this.verticalPadding,
+          horizontal: 25,
         ),
       ),
-      onPressed: this.onPressed,
+      style: ButtonStyle(
+        backgroundColor: MaterialStateProperty.all(color),
+        shape: MaterialStateProperty.all(
+          RoundedRectangleBorder(
+            borderRadius: BorderRadius.circular(BORDER_RADIUS_LARGE),
+            side: BorderSide(
+              color: Color.fromRGBO(40, 40, 40, 1),
+              width: 2,
+            ),
+          ),
+        ),
+      ),
     );
   }
 }
