@@ -2,8 +2,8 @@ package main
 
 import (
 	"github.com/go-flutter-desktop/go-flutter"
-	file_picker "github.com/miguelpruivo/flutter_file_picker/go"
 	"github.com/go-flutter-desktop/plugins/url_launcher"
+	file_picker "github.com/miguelpruivo/flutter_file_picker/go"
 )
 
 var options = []flutter.Option{
