@@ -1,6 +1,5 @@
 import 'package:dynamic_color_theme/dynamic_color_theme.dart';
 import 'package:log_storage_client/utils/constants.dart' as constants;
-import 'package:log_storage_client/utils/constants.dart';
 import 'package:log_storage_client/utils/utils.dart' as utils;
 import 'package:log_storage_client/widgets/app_layout.dart';
 import 'package:flutter/material.dart';
@@ -13,6 +12,7 @@ void main() {
 class MyApp extends StatelessWidget {
   ThemeData _buildTheme(Color accentColor, bool isDark) {
     return ThemeData(
+      dividerColor: constants.DARK_GREY,
       primarySwatch: Colors.blueGrey,
       primaryColor: Color.fromRGBO(32, 33, 37, 1),
       canvasColor: Color.fromRGBO(22, 23, 27, 1),
@@ -68,11 +68,24 @@ class MyApp extends StatelessWidget {
         ),
       ),
       iconTheme: IconThemeData(
-        color: TEXT_COLOR,
+        color: constants.TEXT_COLOR,
       ),
       unselectedWidgetColor: Colors.white,
       floatingActionButtonTheme: FloatingActionButtonThemeData(
         foregroundColor: accentColor,
+      ),
+      tooltipTheme: TooltipThemeData(
+        decoration: BoxDecoration(
+          color: constants.DARK_GREY,
+          borderRadius: BorderRadius.circular(
+            constants.BORDER_RADIUS_SMALL,
+          ),
+        ),
+        textStyle: TextStyle(
+          fontSize: 15,
+          fontWeight: FontWeight.w300,
+          color: constants.TEXT_COLOR,
+        ),
       ),
     );
   }
@@ -86,7 +99,7 @@ class MyApp extends StatelessWidget {
       defaultColor: Color.fromRGBO(1, 176, 117, 1),
       themedWidgetBuilder: (BuildContext context, ThemeData theme) {
         return MaterialApp(
-          title: 'E-Motion Rennteam',
+          title: 'Log Storage Client',
           theme: theme,
           home: AppLayout(),
         );
