@@ -2,13 +2,15 @@ import 'dart:io';
 
 import 'package:log_storage_client/models/storage_connection_credentials.dart';
 import 'package:log_storage_client/models/storage_object.dart';
-import 'package:log_storage_client/utils/app_settings.dart';
 import 'package:log_storage_client/utils/constants.dart';
+import 'package:log_storage_client/utils/i_app_settings.dart';
 import 'package:log_storage_client/utils/locator.dart';
 import 'package:log_storage_client/utils/storage_manager.dart'
     as StorageManager;
 import 'package:log_storage_client/services/progress_service.dart';
 import 'package:log_storage_client/utils/utils.dart';
+import 'package:log_storage_client/utils/storage_object_sorting.dart'
+    as StorageObjectSorting;
 import 'package:log_storage_client/widgets/floating_action_button_position.dart';
 import 'package:log_storage_client/widgets/storage_object_table.dart';
 import 'package:log_storage_client/widgets/storage_object_table_header.dart';
@@ -23,8 +25,9 @@ class RemoteLogFilesView extends StatefulWidget {
 }
 
 class _RemoteLogFilesViewState extends State<RemoteLogFilesView> {
-  List<StorageObject> _storageObjects = new List();
-  List<StorageObject> _selectedStorageObjects = new List();
+  IAppSettings _appSettings = locator<IAppSettings>();
+  List<StorageObject> _storageObjects = [];
+  List<StorageObject> _selectedStorageObjects = [];
   Function _onDownloadFabPressed;
   StorageConnectionCredentials _credentials;
   String _currentDirectory = '/';
@@ -35,7 +38,7 @@ class _RemoteLogFilesViewState extends State<RemoteLogFilesView> {
   @override
   void initState() {
     super.initState();
-    getStorageConnectionCredentials().then((cred) {
+    this._appSettings.getStorageConnectionCredentials().then((cred) {
       this._credentials = cred;
       if (mounted) {
         this._loadStorageObjects();
@@ -55,24 +58,26 @@ class _RemoteLogFilesViewState extends State<RemoteLogFilesView> {
     });
   }
 
-  void _loadStorageObjects() {
+  Future<void> _loadStorageObjects() async {
     setState(() {
       this._storageObjects = null;
       this._allStorageObjectsSelected = false;
       this._failedToListStorageObjects = false;
     });
 
-    StorageManager.listObjectsInRemoteStorage(
-      this._credentials,
-      StorageManager.sortByDirectoriesFirstThenFiles,
-      storagePath: this._currentDirectory,
-    ).then((storageObjects) {
+    try {
+      final storageObjects = await StorageManager.listObjectsInRemoteStorage(
+        this._credentials,
+        StorageObjectSorting.sortByDirectoriesFirstThenFiles,
+        storagePath: this._currentDirectory,
+      );
+
       if (mounted) {
         setState(() {
           this._storageObjects = storageObjects;
         });
       }
-    }).catchError((error) {
+    } catch (error) {
       if (mounted) {
         setState(() {
           this._failedToListStorageObjects = true;
@@ -82,7 +87,7 @@ class _RemoteLogFilesViewState extends State<RemoteLogFilesView> {
       ScaffoldMessenger.of(context).showSnackBar(
         getSnackBar('Failed to list objects. Error: $error', true),
       );
-    });
+    }
   }
 
   /// Changes the currently displayed directory to the given parameter [absolutePath].
